tests/wpt/web-platform-tests/fetch/sec-metadata/appcache.tentative.https.sub.html

<!DOCTYPE html>
<html manifest="/fetch/sec-metadata/resources/record-header.py?file=appcache-manifest{{$id:uuid()}}">
<meta name="timeout" content="long">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/sec-metadata/resources/helper.js></script>
<script src=/common/utils.js></script>
<body></body>
<script>
  async_test(t => {
    window.applicationCache.oncached = window.applicationCache.onnoupdate = window.applicationCache.onerror = t.step_func(e => {
        fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=appcache-manifest{{$id}}")
          .then(t.step_func(response => response.text()))
          .then(t.step_func_done(text => assert_header_equals(text, {
            "dest": "",
            "site": "same-origin",
            "user": "",
            "mode": "no-cors"
          })))
          .catch(t.unreached_func("Fetching and verifying the results should succeed."));
    });
  }, "Appcache!");
</script>
